% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_hex_code.R
\name{generate_colors_to_interpolate}
\alias{generate_colors_to_interpolate}
\title{Generate all possible combination of n colors from the total number N of colors provided. Will be used later for interpolation.}
\usage{
generate_colors_to_interpolate(colors, n)
}
\arguments{
\item{colors}{A tibble with color_names and hex_codes as columns.}

\item{n}{An integer between 1 and number of colors provided.}
}
\value{
A list of length C(N,n), each element containing n colors, where C(N,n) denotes the number of n-combination from a set of N elements.
}
\description{
Generate all possible combination of n colors from the total number N of colors provided. Will be used later for interpolation.
}
\examples{
datafile <- system.file("mycolors.csv", 
                        package = "stadlColorPaletteR")
mycolors <- readr::read_csv(datafile)

# combinations with 2 colors
generate_colors_to_interpolate(mycolors, 2)

#combinations with all colors
generate_colors_to_interpolate(mycolors, nrow(mycolors))

}
