% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_hex_code.R
\name{get_color_hex_code}
\alias{get_color_hex_code}
\title{Get hex code of colors}
\usage{
get_color_hex_code(..., colors = mycolors)
}
\arguments{
\item{...}{Names of colors as character}

\item{colors}{A tibble with color_names and hex_codes as columns}
}
\value{
Hex code of colors specified
}
\description{
Get hex code of colors
}
\examples{
my_median(1:12)

# Example with your dataset in "inst/"
datafile <- system.file("nyc_squirrels_sample.csv", package = "stadlColorPaletteR")
nyc_squirrels <- read.csv(datafile, encoding = "UTF-8")
# Apply my function
my_median(nyc_squirrels[,"hectare_squirrel_number"])
}
