% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_stadl.R
\name{scale_color_stadl}
\alias{scale_color_stadl}
\title{Create function for color scales}
\usage{
scale_color_stadl(
  colors_to_interpolate = "pal1",
  base_colors_set = stadl2,
  discrete = TRUE,
  reverse = FALSE,
  subset = as.vector(NULL),
  ...
)
}
\arguments{
\item{colors_to_interpolate}{Name of the colors to interpolate. It is the name of a list element as generated by the function \code{\link[=generate_colors_to_interpolate]{generate_colors_to_interpolate()}}.}

\item{discrete}{Boolean indicating whetehr or not to use discrete scale}

\item{reverse}{Boolean indicating whether or not to reverse the order of the colors.}

\item{...}{Other arguments to \link[grDevices:colorRamp]{grDevices::colorRampPalette}}

\item{base_colors}{A list of colors to interplotate. Output from \code{\link[=generate_colors_to_interpolate]{generate_colors_to_interpolate()}}.}
}
\value{
A function for scales to use in ggplot.
}
\description{
Create function for color scales
}
\examples{
require(ggplot2)
datafile <- system.file("mycolors.csv", 
                        package = "stadlColorPaletteR")
mycolors <- readr::read_csv(datafile)
ggplot(iris, 
       aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 2, alpha = 1) +
  scale_color_stadl('pal1', generate_colors_to_interpolate(mycolors,3)) +
  ggtitle("stadl")+theme_minimal()

}
