% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_stadl.R
\name{scale_fill_stadl}
\alias{scale_fill_stadl}
\title{Create function for filling scales.}
\usage{
scale_fill_stadl(
  colors_to_interpolate = "pal1",
  base_colors_set = stadl2,
  discrete = TRUE,
  reverse = FALSE,
  subset = as.vector(NULL),
  ...
)
}
\arguments{
\item{colors_to_interpolate}{Name of the colors to interpolate. It is the name of a list element as generated by the function \code{\link[=generate_colors_to_interpolate]{generate_colors_to_interpolate()}}.}

\item{reverse}{Boolean indicating whether or not to reverse the order of the colors.}

\item{...}{Other arguments to \link[grDevices:colorRamp]{grDevices::colorRampPalette}}

\item{base_colors}{A list of colors to interplotate. Output from \code{\link[=generate_colors_to_interpolate]{generate_colors_to_interpolate()}}.}
}
\value{
A function interpolating colors.
}
\description{
Create function for filling scales.
}
