% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadlColorPaletteR-package.R
\docType{package}
\name{stadlColorPaletteR-package}
\alias{stadlColorPaletteR}
\alias{stadlColorPaletteR-package}
\title{stadlColorPaletteR: STADL Color Palettes}
\description{
Build color palettes based on the five basic STADL colors.
}
\author{
\strong{Maintainer}: Thomas Thomas \email{thomas.gsponer@stadl-partners.ch} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
