% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_hex_code.R
\name{get_color_hex_code}
\alias{get_color_hex_code}
\title{Get hex code of colors}
\usage{
get_color_hex_code(..., base_colors)
}
\arguments{
\item{...}{Names of colors as character}

\item{base_colors}{A tibble with color_names and hex_codes as columns}
}
\value{
Hex code of colors specified
}
\description{
Get hex code of colors
}
\examples{
datafile <- system.file("mycolors.csv", 
                        package = "stadlColorPaletteR")
mycolors <- readr::read_csv(datafile)
get_color_hex_code('palevioletred2', base_colors = mycolors)

}
